Refactor label_generators into a submodule
