+++ src/effiara/label_generators/__init__.py
@@ -1,0 +1,1 @@
+from .label_generator import LabelGenerator, DefaultLabelGenerator
